+++ HISTORY.rst
@@ -163,7 +163,9 @@
   into a Mayan app. With change adds the new settings:
   "COMMON_AUTOADMIN_EMAIL", "AUTOADMIN_PASSWORD", and
   "AUTOADMIN_USERNAME".
-
+- Changed the use of the list/zip combinarion to generate
+  the full list of metadata lookup choices to a generator.
+  
 3.1.9 (2018-11-01)
 ==================
 - Convert the furl instance to text to allow serializing it into